The main part of the task is completed

+++ ProfileBook/ProfileBook/ViewModels/MainListViewModel.cs
@@ -1,13 +1,12 @@
-﻿using Prism.Mvvm;
+﻿using Acr.UserDialogs;
+using Prism.Mvvm;
 using Prism.Navigation;
 using ProfileBook.Models;
 using ProfileBook.Services.AuthorizationServices;
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using Xamarin.Forms;
@@ -22,6 +21,7 @@ namespace ProfileBook.ViewModels
         private ICommand _exitCommand;
         private ICommand _addProfileCommand;
         private ICommand _deleteProfileCommand;
+        private ICommand _editProfileCommand;
         private readonly IAuthorizationService _authorization;
         public ObservableCollection<Profile> ProfileCollection { get; set; }
 
@@ -33,11 +33,9 @@ namespace ProfileBook.ViewModels
         }
 
         public void InitTable()
-        {
-            
+        {            
             ProfileCollection = new ObservableCollection<Profile>(App.DatabaseProfile.GetItems().Where(p => p.User_Id==App.UserLogin));
             LableText = ProfileCollection.Count == 0 ? "No profiles added" : "";
-
         }
         public ICommand ExitCommand => _exitCommand ?? (_exitCommand = new Command(
                         async () => await ExitFromProfileAsync())
@@ -48,31 +46,53 @@ namespace ProfileBook.ViewModels
         public ICommand DeleteProfileCommand => _deleteProfileCommand ?? (_deleteProfileCommand = new Command(
                         async (Object obj) => await DeleteProfile(obj))
                         );
+        public ICommand EditProfileCommand => _editProfileCommand ?? (_editProfileCommand = new Command(
+                        async (Object obj) => await EditProfile(obj))
+                        );
 
+        async Task EditProfile(object obj)
+        {
+            _profile = obj as Profile;
+            var parametr = new NavigationParameters
+            {
+                {"profileImage", _profile.ProfileImage },
+                {"nickName", _profile.NickName },
+                {"name", _profile.Name },
+                {"description", _profile.Description },
+                {"dateTime", _profile.DateTimePr },
+                {"profileId", _profile.Id }
+            };
+            await _navigationService.NavigateAsync(new Uri("AddEditProfileView", UriKind.Relative), parametr);
+        }
         async Task DeleteProfile(object obj)
         {
             _profile = obj as Profile;
             int profileId;
-            if (_profile != null)
+            var result = await UserDialogs.Instance.ConfirmAsync(new ConfirmConfig
+            {
+                Message = "Do you confirm deletion?",
+                OkText = "Delete",
+                CancelText = "Cancel"
+            });
+            if (_profile != null && result == true)
             {
                 profileId = _profile.Id;
                 App.DatabaseProfile.DeleteItem(profileId);
                 ProfileCollection = new ObservableCollection<Profile>(App.DatabaseProfile.GetItems().Where(p => p.User_Id == App.UserLogin));
             }
-            else throw new ArgumentException("Fuck!");
             await _navigationService.NavigateAsync(new System.Uri("http://www.ProfileBook/NavigationPage/MainListView", System.UriKind.Absolute));
         }
 
         async Task AddProfileAsync()
         {
-            await _navigationService.NavigateAsync("AddEditProfileView");
+            await _navigationService.NavigateAsync(new Uri("AddEditProfileView", UriKind.Relative));
         }
 
         async Task ExitFromProfileAsync()
         {
             _authorization.AddUodateUserId(-1);
             _authorization.ToWriteLoginId();
-            await _navigationService.NavigateAsync(new System.Uri("http://www.ProfileBook/SignInPageView", System.UriKind.Absolute));
+            await _navigationService.NavigateAsync(new System.Uri("http://www.ProfileBook/NavigationPage/SignInPageView", System.UriKind.Absolute));
         }
         public string LableText
         { 
@@ -82,8 +102,6 @@ namespace ProfileBook.ViewModels
         public bool CanNavigate(INavigationParameters parameters)
         {
             return true;
-        }
-
-       
+        }       
     }
 }
